requirements.txt
easyocr
PyMuPDF 
Pillow
ollama
customtkinter